added error table for supplemental data reporting errors. error are now reported for each supplemental data (Race, Sex, Ethnicity, Payer) with missmatched expected and reported values.
